+++ src/main/java/com/rogerhagerbaum/parkerning/service/PersonService.java
@@ -1,5 +1,6 @@
 package com.rogerhagerbaum.parkerning.service;
 
+import com.rogerhagerbaum.parkerning.exceptionHandling.EntityException;
 import com.rogerhagerbaum.parkerning.module.dto.CarDto;
 import com.rogerhagerbaum.parkerning.module.dto.PersonDto;
 import com.rogerhagerbaum.parkerning.module.entity.Car;
@@ -8,6 +9,7 @@ import com.rogerhagerbaum.parkerning.repositories.PersonRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,7 +25,6 @@ public class PersonService {
     public PersonDto addPerson(PersonDto personDto) {
         Person createNew = modelMapper.map(personDto, Person.class);
         personRepository.save(createNew);
-
         return modelMapper.map(createNew, PersonDto.class);
     }
     public List<PersonDto> getAll(){
@@ -33,12 +34,18 @@ public class PersonService {
                 .map(PersonDto::new)
                 .collect(Collectors.toList());
     }
-    public PersonDto getById(int id){
+    public PersonDto getById(Long id){
         Person person = personRepository.findPersonById(id);
+        if (person == null){
+            throw new EntityException(String.format("Person: %d not found", id), HttpStatus.NOT_FOUND);
+        }
         return modelMapper.map(person , PersonDto.class);
     }
-    public PersonDto deletePerson(int id) {
+    public PersonDto deletePerson(Long id) {
         Person person = personRepository.findPersonById(id);
+        if (person == null){
+            throw new EntityException(String.format("Person: %d not found", id), HttpStatus.NOT_FOUND);
+        }
         personRepository.delete(person);
         return modelMapper.map(person, PersonDto.class);
     }
